StripedLock fix for Integer.MIN_VALUE hashcode overflow
This is the same as reported in auroraArc: https://github.com/threadly/auroraArc/issues/11
Please see that issue for discussion on this solution.

+++ src/main/java/org/threadly/concurrent/lock/StripedLock.java
@@ -63,6 +63,6 @@ public class StripedLock {
    * @return consistent Object for a given hash code
    */
   public Object getLock(int hashCode) {
-    return locks[Math.abs(hashCode) % locks.length];
+    return locks[(int)(Math.abs((long)hashCode) % locks.length)];
   }
 }
